requirements.txt
pyyaml~=5.4.0
boto3~=1.17.0
black